fix log rotation: add explicit cast for files_to_delete
And reduce the variable size from 64 bits to ssize_t, as it is enough.

further info: If the subtraction between file_map.size() and _max_files is negative the result will
still be treated as unsigned and therefore become a large positive number before casting it to
signed.

+++ src/mavlink-router/logendpoint.cpp
@@ -194,9 +194,9 @@ void LogEndpoint::_delete_old_logs()
     int64_t bytes_to_delete = _min_free_space - free_space;
     // If the configured value for _max_files is 0, then set this to -1 to indicate that we've
     // already deleted enough files.
-    int64_t files_to_delete = _max_files > 0 ? file_map.size() - _max_files : -1;
+    ssize_t files_to_delete = _max_files > 0 ? (ssize_t)file_map.size() - _max_files : -1;
 
-    log_debug("[Log Deletion] Files to delete: %ld", files_to_delete);
+    log_debug("[Log Deletion] Files to delete: %zd", files_to_delete);
 
     // Delete the logs in order until there's enough free space, and few enough files
     // It is possible for this loop to run only once and return immediately, if we don't actually
